+++ src/main/java/org/springframework/samples/petclinic/system/WelcomeController.java
@@ -1,0 +1,14 @@
+package org.springframework.samples.petclinic.system;
+
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+class WelcomeController {
+
+    @RequestMapping("/")
+    public String welcome() {
+        return "welcome";
+    }
+}
