--- src/main/java/org/springframework/samples/petclinic/web/WelcomeController.java
@@ -1,14 +1,0 @@
-package org.springframework.samples.petclinic.web;
-
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-public class WelcomeController {
-
-    @RequestMapping("/")
-    public String welcome() {
-        return "welcome";
-    }
-}
