--- src/main/java/org/springframework/samples/petclinic/web/package-info.java
@@ -1,5 +1,0 @@
-/**
- * The classes in this package represent PetClinic's web presentation layer.
- */
-package org.springframework.samples.petclinic.web;
-
