+++ src/main/resources/templates/pets/createOrUpdateVisitForm.html
@@ -19,12 +19,12 @@
       </tr>
     </thead>
     <tr>
-      <td th:text="${visit.pet.name}" /></td>
+      <td th:text="${pet.name}" /></td>
       <td
-        th:text="${#calendars.format(visit.pet.birthDate, 'yyyy-MM-dd')}" /></td>
-      <td th:text="${visit.pet.type}" /></td>
+        th:text="${#calendars.format(pet.birthDate, 'yyyy-MM-dd')}" /></td>
+      <td th:text="${pet.type}" /></td>
       <td
-        th:text="${visit.pet.owner?.firstName + ' ' + visit.pet.owner?.lastName}" /></td>
+        th:text="${pet.owner?.firstName + ' ' + pet.owner?.lastName}" /></td>
     </tr>
   </table>
 
@@ -38,7 +38,7 @@
 
     <div class="form-group">
       <div class="col-sm-offset-2 col-sm-10">
-        <input type="hidden" name="petId" th:value="${visit.pet.id}" />
+        <input type="hidden" name="petId" th:value="${pet.id}" />
         <button class="btn btn-default" type="submit">Add Visit</button>
       </div>
     </div>
@@ -51,7 +51,7 @@
       <th>Date</th>
       <th>Description</th>
     </tr>
-    <tr th:if="${!visit['new']}" th:each="visit : ${visit.pet.visits}">
+    <tr th:if="${!visit['new']}" th:each="visit : ${pet.visits}">
       <td th:text="${#calendars.format(visit.date, 'yyyy-MM-dd')}" /></td>
       <td th:text=" ${visit.description}" /></td>
     </tr>
